refactoring core module
