+++ pulumi/core/utils.py
@@ -1,10 +1,23 @@
-# ./pulumi/core/utils.py
-# Description: Utility functions for sanitizing compliance metadata labels.
+# pulumi/core/utils.py
+
+"""
+Utility Functions Module
+
+This module provides generic, reusable utility functions.
+It includes resource transformations, Helm interactions, and miscellaneous helpers.
+"""
 
 import re
 import pulumi
 import pulumi_kubernetes as k8s
 from typing import Optional, Dict, Any
+import requests
+import logging
+import yaml
+from packaging.version import parse as parse_version, InvalidVersion, Version
+
+# Set up basic logging
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
 def set_resource_metadata(metadata: Any, global_labels: Dict[str, str], global_annotations: Dict[str, str]):
     """
@@ -42,21 +55,54 @@ def generate_global_transformations(global_labels: Dict[str, str], global_annota
 
     pulumi.runtime.register_stack_transformation(global_transform)
 
-def sanitize_label_value(value: str) -> str:
+def get_latest_helm_chart_version(url: str, chart_name: str) -> str:
     """
-    Sanitizes a label value to comply with Kubernetes naming conventions.
+    Fetches the latest stable version of a Helm chart from the given URL.
 
     Args:
-        value (str): The value to sanitize.
+        url (str): The URL of the Helm repository index.
+        chart_name (str): The name of the Helm chart.
 
     Returns:
-        str: The sanitized value.
+        str: The latest stable version of the chart.
     """
-    value = value.lower()
-    sanitized = re.sub(r'[^a-z0-9_.-]', '-', value)
-    sanitized = re.sub(r'^[^a-z0-9]+', '', sanitized)
-    sanitized = re.sub(r'[^a-z0-9]+$', '', sanitized)
-    return sanitized[:63]
+    try:
+        logging.info(f"Fetching URL: {url}")
+        response = requests.get(url)
+        response.raise_for_status()
+
+        index = yaml.safe_load(response.content)
+        if chart_name in index['entries']:
+            chart_versions = index['entries'][chart_name]
+            stable_versions = [v for v in chart_versions if is_stable_version(v['version'])]
+            if not stable_versions:
+                logging.info(f"No stable versions found for chart '{chart_name}'.")
+                return "Chart not found"
+            latest_chart = max(stable_versions, key=lambda x: parse_version(x['version']))
+            return latest_chart['version']
+        else:
+            logging.info(f"No chart named '{chart_name}' found in repository.")
+            return "Chart not found"
+
+    except requests.RequestException as e:
+        logging.error(f"Error fetching data: {e}")
+        return f"Error fetching data: {e}"
+
+def is_stable_version(version_str: str) -> bool:
+    """
+    Determines if a version string represents a stable version.
+
+    Args:
+        version_str (str): The version string to check.
+
+    Returns:
+        bool: True if the version is stable, False otherwise.
+    """
+    try:
+        parsed_version = parse_version(version_str)
+        return isinstance(parsed_version, Version) and not parsed_version.is_prerelease and not parsed_version.is_devrelease
+    except InvalidVersion:
+        return False
 
 def extract_repo_name(remote_url: str) -> str:
     """
